requirements.txt
Django==5.1.6
djangorestframework==3.15.2
drf_spectacular==0.28.0
minio==7.2.15
django-cors-headers